Migrate summary data as 32 bit float

+++ tests/integration_tests/test_storage_migration.py
@@ -138,7 +138,7 @@ def test_that_storage_matches(
             tuple(ensemble.get_realization_list_with_responses("summary")),
         )
         snapshot.assert_match(
-            summary_data.to_dataframe().astype(np.float32).transform(np.sort).to_csv(),
+            summary_data.to_dataframe().transform(np.sort).to_csv(),
             "summary_data",
         )
         snapshot.assert_match_dir(
